fixed the wrong pressure

+++ src/datalogger.py
@@ -113,7 +113,7 @@ class fileDataLogger:
                 if nodeId == 2: # temp_out, pressure, humidity
                     temp = float(values[fields.index("temp")]) if "temp" in fields else 999
                     pressure = float(values[fields.index("pressure")]) if "pressure" in fields else 0
-                    humidity = float(values[fields.index("humidity")]) if "humidity" in fields else -1
+                    humidity = int(values[fields.index("humidity")]) if "humidity" in fields else -1
                     if temp > -10 and temp < 40:
                         new_data["temp_in"] = new_data["temp_out"] = temp
                     if pressure > 900 and pressure < 1100:
